✨(ashley) add support for unique LTI launch URL per forum
We want to have a unique LTI launch URL per forum, and to be able to
have multiple forum per LTI context. As a result, each forum is now
identified by an UUID instead of the lti_context that created the
forum. A new model (LTIContext) has been introduced in order to
materialize a LTI context and to link it to a forum. In addition,
LTIContext has the responsability to synchronize the group membership
of a user inside its context, in replacement of the
ashley.permissions.groups module.