✨(ashley) render emojis with emojione
Emojis are rendered with emojione in ashley WISYWIG editor. But when a
forum post is displayed, emojis are rendered with the native font of
the user's browser. In consequence, the rendering is inconsistent and
sometime emojis are not displayed at all. To improve that, emojis are
now rendered via emojione on the forum posts.