✨(ashley) add support for Moodle in ashley's LTI authentication backend
The LTI parameter containing the remote user id depends on the LTI
consumer. As a result, when using Ashley on Moodle, it fails with the
following error : "Unable to find param lis_person_sourcedid in LTI
request". This commit adds support for Moodle, which use the LTI
parameter `ext_user_username` to send the remote user id.