💥(lti) replace lti_provider app with the `django-lti-toolbox` dependency
We want to use `django-lti-toolbox` dependency instead of the
`lti_provider` app included in this project. This is a copy of this
app, it has been externalized in order to use it in multpile projects.

BREAKING CHANGE: The migration part of this refactoring is tricky.
If you upgrade from a previous ashley version, you MUST execute the
instructions documented in UPGRADE.md