✨(forum) add rename feature
When a forum is created, its name is generated with the course name
contained in the LTI payload. But sometimes, the course name is
unknown. As a result, the forum name is just an ID, this is not very
user friendly. We would like forum administrators to be able to rename
their forums.