✨(back) add management command `sync_group_permissions`
Ashley permissions for LTI groups are defined with the setting
ASHLEY_DEFAULT_FORUM_ROLES_PERMISSIONS. If the set of default
permissions is modified, it is only applied to the groups that are
created after the setting update and it can lead to inconsistent
behavious between forums.

The `sync_group_permission` management command allows to synchronize
your group permissions in database with the set of permission defined
in this setting.