🐛(front) fix Draft.js editor resize issue with static toolbar
We used a CSS hack to be able to place the Draft.js static toolbar
plugin above the editor, because it was not possible to place directly
the static-toolbar component before the editor.

This hack does not handle well the resizing of the editor, because of
the way it computes the height of the editor. It is one time behind
the reality.