✨(auth) make public username and email optional in the LTI payload
In some contexts (i.e. when Ashley is integrated in the Open edX
studio), the user email and the public username are not provided in
the LTI payload. As a result, a blank page is displayed instead of the
forum. This is because of the LTI authentication backend provided with
ashley, which require these parameters.