🔒(auth) generate a unique user id for openedx studio users
When OpenedX studio initiates a LTI request to render ashley in a
xblock, it always sends "student" as the user_id. This is a problem,
because all openedX studio users of the same LTI consumer share the
same user account on ashley. And as a result, they have an instructor
access to every forums related to their LTI consumer.

To solve this problem, we now detect openedx studio users and generate
a unique user id specific to their LTI context id.