🐛(search) fix serialization error in search index update
The haystack command `update_index` crashes while indexing posts with
a serialization error. It happens when a message is posted by a user
having an empty public_username field.

This is due to the fact that in this specific case, we are returning a
default public_username ("Anonymous") translated with gettext_lazy.
And lazy proxy values are not directly serializable into JSON.

We have chosen not to index the default "Anonymous" username.