✨(back) limit forum listing to the current user LTIContext id
When a user is connected to Ashley via LTI, he can see and interact
with all forums he has access to, based on its group membership.

It can be confusing for the user, because he comes from a specific LTI
context and he does not expect to see information related to other
LTI contexts.

In addition to that, group membership synchronization is made during
the LTI authentication process. And it only synchronize groups for the
current LTI context. So, if we let users access to resources related
to other LTI contexts, we cannot guarantee that user permissions are
correct and it can lead to security issues.

That's why we decided to limit the user to the scope of its current
LTI context.