✨(back) allow identical LTI launch URL for two distinct forums
In OpenEdx when we duplicate a course, we automatically copy the xblock for
Ashley. Even if we have a new LTIContext id as it’s a new course, the forum
will have the same LTI launch URL. We want to create a new forum for each LTI
context. This way, we won't have to modify the LTI launch URL of all the
Ashley xblocks in the course to create new empty forums.