⚡️(back) add index to the Forum model
In commit `f8943b5a1d4ccd8823fd204791b4f7a2d5343277` we removed  the unique
property of the field `lti_id`. As a result, the index automatically created
with the unique property got removed. This field is used by `ForumLTIView` to
find the Forum and it’s essential in terms of performance that we maintain an
index property on this field. For this purpose, we've added one explicitly.