✨(forum) add mentions in editor for active users in current topic
A nice feature in a forum is to be able to mention someone specific using
the `@` keyword. It enables to target a person directly and to fluidly
exchanges. We’ve added a draft-js plugin to do this task and propose
to mention any user that has been active in the current topic.