✨(forum) sort discussion topics
The view listing the different topics in the forum is by default ordered by
last messages which enables to see recent activity. If we want to see which
subjects in the past were dynamic, it’s interesting to be able to sort
discussions from other parameters such as number of replies or views. On the
other hand, if we have a lot of history, we still want to focus on the latest
posts, so it’s interesting to combine sorting fields. For this purpose, this
feature introduces sorting by multiple columns. This source code is inspired
and done the same way it’s done in Django’s admin.

Resolve #66