✨(forum) sort discussion topics
The view listing the different topics in the forum is by default ordered by
last messages which enables to see recent activity. If we want to see which
subjects in the past were dynamic, it’s interesting to be able to sort
discussions from other parameters such as number of replies or views.

Resolve #66