✨(back) add role moderator and permission to manage moderators
Instructors might need help on administrating forums for a given LTI Context.
On each forum creation a new moderator group specific to this forum is created.
This group has whole instructor permissions group. We add a new permission
`can_manage_moderator` and add it to instructor’s group. This permission will
be used when the feature will be tuned. Moderator group is an internal group
and won’t get deleted when user is connecting with a LTI role student or
instructor.