✨(front) add react components autoload and internationalisation
To build a modern and dynamic interface, Ashley will add on frontend features
and react components. We want a system where components can be easily and
automatically added to our view. As Ashley is international and it's backend
already translated, we want to add the internationalisation feature for our
components.