✨(front) add autoload and internationalization for AshleyEditor
As AshleyEditor component gets more and more features, it will be easier to
add and maintain translations directly from the frontend.  Autoload and
internationalization of react component has been added in the previous commit
b0deca212373a8e6dff20de578a387b448093e1e, so we use this system to add
AshleyEditor component to the view.