✨(back) add DRF and Amazon S3 to upload image
Images can take a lot of space and resources. We want to use an external
service for this job. We configure Amazon S3 in the project. To easily upload
images from the frontend, we add a new endpoint that will associate images,
forum, and user in a new model UploadImage. This way, we will be able on
cleaning forums or deleting users to clean up our image's repository on Amazon