🔧(front) upgrade node version
Docker-compose uses the last node version : node 14. Circle and docker config
files are using the version 10 of node. Node 10 has no compatibility anymore
with some packages. To keep the project updated and to have all the versions
homogenized, an upgrade of node version to version 14 is needed in config files.