✨(back) allow instructors to archive a forum
If a forum is created by mistake by an instructor on Ashley, it cannot
be deleted. And as a result, it is displayed to all students coming
from the same LTI context.

This commit introduces a "delete" action on the forum. It is rather an
archiving of the forum than a deletion: the forum is flagged as
archived and not accessible anymore.