🔧(setup.cfg) target elasticsearch version with PEP 440 norm
Renovate service auto updates python packages used in Ashley depending on
the setup.cfg file. The way it's actually written in this config file,
doesn't constrain elasticsearch's version between v5 and v6. Ashley is
meant to work with the latest release of version 5 of elasticsearch. We
target directly this version by using the pattern 5.*.