🔧(renovate) ignore js dependencies bootstrap
Renovate service is trying to upgrade the version of bootstrap. Ashley is not
configured yet to properly work with version 5 of bootstrap. There are
different breaking styles that need to be fixed first. We still want
renovate to work for other dependencies, so we explicitly exclude
bootstrap from the config of renovate.