🐛(forum) fix sorting on sticky and announcements topics
There are different kind of topics: regular, sticky and announcements topics.
Sticky topics are supposed to stay on top of the topic listing. Announcements
are displayed in another block that is above the regular topics block and has
its own header columns. Previous commit 3f92f4a enabled sorting topics on
different header columns. This new feature introduces two different
inconstancies. First, sticky topics were sorted as well whereas they are meant
to stay on top of the listing. Second, columns of announcements block when the
order was requested in URL showed toggling arrow as if they were sorted as well.
We only want to sort regular topic discussions and not announcements.
Announcements must stay with the default order which is ordered on the creation
date. The commit introduces these two fixes.