✨(forum) allow to move topics in the same LTIContext
A feature that has been requested by teachers to moderate forums is to be able
to move topics. Django Machina already has this feature ready. We upgrade the
moderator group permissions on assigning the permission ’can_move_topics’.
Moderators and instructors will now be able to move topics from one forum to
another one as long as it's part of the same LTIContext.