🔧(sandbox) allowing multiple forums for a same LTIContext on sandbox
In development mode, the sandbox allows using Ashley via a dummy LTI consumer.
So far, we can’t create several forums for the same course, since the forum
LTI id is generated from the course ID. This prevents us from testing features
that we are trying to implement. Enabling to set the LTI id value will fix this
limit.