✨(xapi) track topic views with XAPI events
We want to track our users' actions using the XAPI standard and
collect these events for analysis.

For this purpose, a xapi utility module has been added to the ashley
application to help building API statements, it is based on the tincan
library.

The events XAPI are emitted on a specific logger, per LTI consumer.

The tracking of a first event (when a user view a topic) has been
implemented in this PR.