✨(auth) set a default public username if empty for admin/instructor
When the public username is not provided in the LTI payload, we ask
the user to define the username he would like to use. In some
contexts (i.e. when Ashley is integrated into the Open edX studio),
the public username will then be set for the entire team. In the
external tool loading Ashley, this is confusing. We have decided for
users that have administrator or instructor roles, to force a default
username depending on the role. This will avoid the situation when an
instructor or an administrator was thinking he was setting a username
only for himself.