✨(forum) use the name of a previous forum with the same `lti_id`
When duplicating courses from an LMS, forums get copied. By default,
each forum has a name set with the value of the course's name. It's
more likely, when a course gets copied, that instructors want to keep
the same forum's names. To simplify this use case, we define the
default name of each forum with an existing `lti_id` to the previous
name defined for the other `context.id`.

Resolve https://github.com/openfun/ashley/issues/231