🐛(auth) set a default public username for existing users
In previous commit 06bfbc3, we defined a default public username for
instructors and administrators that had none defined (i.e. when Ashley
is integrated into the Open edX studio). No change had been done
for users that were already set and had an empty public_username, this is now
implemented.