⚡️(ashley) stop calculating the list of roles for a user
The sndpoint that lists users that are part of a group was
calculated for each user the list of the role he was
part of. This implied extra queries for each user of the
list. Ashley can be used with forums that have a lot of
participants. Adding systematically this information on
querying listing is not necessary and consumes resources
that can be saved.