🐛(back) search queries restricted on LTIContext
If a user was part of multiple courses, search queries were
taking a lot of processing time. DjangoMachina checks the
permissions of the user for each forum he has access to. We only
need to control that the user has the permission to read the forums
part of the current LTIContext. By limiting this query, we fix
at the same time, the problem of the processing time.
With this new method overridden from DjangoMachina, we exclude as
well archived forums.