✨(forum) search links target directly the refered post
A topic can have multiple posts and a lot of pages. Search
results were targeting the topic on page 1 for any match.
If the post matching the search was on page 10, it made it
hard to relate the search and the post. In this condition,
the results didn't seem accurate. To be more precise, we
now build a link that directly targets the post on the right
page in the topic.

Resolves https://github.com/openfun/ashley/issues/244