👷(checks) add check in CI to ensure all versions are in sync
The current ashley version is mentioned in several places in the code:
backend, frontend, changelog... All these versions should always be
kept in sync. It is not rare to forget to bump a version number when
making a release so this should be checked by the CI before publishing
packages to registries.