✨(forum) unlock all forums of an entire course
A recent feature enabled administrators to lock the forums of a
course. Use cases have shown that sometimes, forums of the
course need to be unlocked. This is the case, for instance, when a
course is duplicated, the forums get copied but they keep the
lock as well. It will now be possible to unlock forums once they
have been locked.