🔥(forum) delete code-editor plugin for draft-js
Historically code-editor plugin was added to ease the
edition of programming language in a code-block in draft-js :
 - the use of the`tab` button was keeping the user in
the draft-js code-block
- an extra bloc was added once we were in the code-block
so the user could exit it without adding one himself.
Unfortunately, using other plugins that need to use the
keyboard event, it appears to be in conflict.
As a matter of fact, adding an extra block is not so
important, as user can still exit the block using the
back arrow key and the use of the tab is not that wise
as it's not a good manner for accessibility, so we finally
decided to just delete the use of this fork and keep the
normal behavior of draft-js.