🐛(back) fix exporting atomic blocks for images
For atomic blocks, we needed a specific behavior for the ones
dealing with the latex formulas. This introduced a problem with
rendering the images. This commit fixes this. To prevent this
kind of error in the future, tests have been added rendering the HTML
using draftjs_exporter. So far only specific decorators were tested
but not all the rendering of multiple draftjs blocks.